+++ SharePointFramework/PortfolioWebParts/src/components/ResourceAllocation/IResourceAllocationProps.ts
@@ -1,4 +1,5 @@
 import { IBaseComponentProps } from '../IBaseComponentProps';
+import * as moment from 'moment';
 
 export interface IResourceAllocationProps extends IBaseComponentProps {
    /**
@@ -15,4 +16,14 @@ export interface IResourceAllocationProps extends IBaseComponentProps {
     * Background color for absence items
     */
    itemAbsenceBgColor?: string;
+
+   /**
+    * Default time start
+    */
+   defaultTimeStart?: [number,  moment.unitOfTime.DurationConstructor];
+
+   /**
+    * Default time end
+    */
+   defaultTimeEnd?: [number, moment.unitOfTime.DurationConstructor];
 }